docs: getRandomNumbers 를 Computer 메서드로 옮기고 utils 폴더 삭제
