+++ src/models/Computer.js
@@ -1,4 +1,5 @@
-import { getRandomNumbers } from '../utils/randomNumber.js';
+import { MissionUtils } from '@woowacourse/mission-utils';
+import { RULE } from '../constants/index.js';
 
 class Computer {
   /**
@@ -9,8 +10,19 @@ class Computer {
   constructor() {
     this.#setNumbers();
   }
+  #getRandomNumbers() {
+    let array = [];
+    while (array.length < RULE.lengthOfNumbers) {
+      const { start, end } = RULE.rangeOfNumber;
+      const number = MissionUtils.Random.pickNumberInRange(start, end);
+      if (!array.includes(number) && number) {
+        array.push(number);
+      }
+    }
+    return array;
+  }
   #setNumbers() {
-    const numbers = getRandomNumbers();
+    const numbers = this.#getRandomNumbers();
     this.#numbers = numbers;
   }
